requirements.txt
Django==3.1.3
Pillow==8.0.1
djangorestframework==3.12.1
django-ckeditor==6.0.0
django-filter==2.4.0
djangorestframework-simplejwt==4.5.0
